minijust-ai/requirements.txt
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
pymilvus>=2.3.0
milvus>=2.3.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0
PyPDF2>=3.0.0
python-docx>=1.1.0
docx2txt>=0.8
PyMuPDF>=1.23.0
langid>=1.1.6
